debug rest
