+++ Identity_Service/nbproject/project.properties
@@ -20,7 +20,7 @@ build.test.classes.dir=${build.dir}/test/classes
 build.test.results.dir=${build.dir}/test/results
 build.web.dir=${build.dir}/web
 build.web.excludes=${build.classes.excludes}
-client.urlPart=
+client.urlPart=/webresources/wbd.identity.token
 compile.jsps=false
 conf.dir=${source.root}/conf
 debug.classpath=${build.classes.dir}:${javac.classpath}
