+++ Identity_Service/setup/sun-resources.xml
@@ -1,23 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE resources PUBLIC "-//Sun Microsystems, Inc.//DTD Application Server 9.0 Resource Definitions //EN" "http://www.sun.com/software/appserver/dtds/sun-resources_1_3.dtd">
 <resources>
-  <jdbc-resource enabled="true" jndi-name="jdbc/Identity_Service" object-type="user" pool-name="Identity_Service">
+  <jdbc-resource enabled="true" jndi-name="wbd2" object-type="user" pool-name="wbdPool">
     <description/>
   </jdbc-resource>
-  <jdbc-resource enabled="true" jndi-name="jdbc/IS" object-type="user" pool-name="Identity_Service">
-    <description/>
-  </jdbc-resource>
-  <jdbc-resource enabled="true" jndi-name="jdbc/wbd2" object-type="user" pool-name="wbdPool">
-    <description/>
-  </jdbc-resource>
-  <jdbc-connection-pool allow-non-component-callers="false" associate-with-thread="false" connection-creation-retry-attempts="0" connection-creation-retry-interval-in-seconds="10" connection-leak-reclaim="false" connection-leak-timeout-in-seconds="0" connection-validation-method="auto-commit" datasource-classname="com.mysql.jdbc.jdbc2.optional.MysqlDataSource" fail-all-connections="false" idle-timeout-in-seconds="300" is-connection-validation-required="false" is-isolation-level-guaranteed="true" lazy-connection-association="false" lazy-connection-enlistment="false" match-connections="false" max-connection-usage-count="0" max-pool-size="32" max-wait-time-in-millis="60000" name="Identity_Service" non-transactional-connections="false" pool-resize-quantity="2" res-type="javax.sql.DataSource" statement-timeout-in-seconds="-1" steady-pool-size="8" validate-atmost-once-period-in-seconds="0" wrap-jdbc-objects="false">
-    <property name="URL" value="jdbc:mysql://localhost:3306/mysql?zeroDateTimeBehavior=convertToNull"/>
-    <property name="User" value="root"/>
-    <property name="Password" value=""/>
-  </jdbc-connection-pool>
   <jdbc-connection-pool allow-non-component-callers="false" associate-with-thread="false" connection-creation-retry-attempts="0" connection-creation-retry-interval-in-seconds="10" connection-leak-reclaim="false" connection-leak-timeout-in-seconds="0" connection-validation-method="auto-commit" datasource-classname="com.mysql.jdbc.jdbc2.optional.MysqlDataSource" fail-all-connections="false" idle-timeout-in-seconds="300" is-connection-validation-required="false" is-isolation-level-guaranteed="true" lazy-connection-association="false" lazy-connection-enlistment="false" match-connections="false" max-connection-usage-count="0" max-pool-size="32" max-wait-time-in-millis="60000" name="wbdPool" non-transactional-connections="false" pool-resize-quantity="2" res-type="javax.sql.DataSource" statement-timeout-in-seconds="-1" steady-pool-size="8" validate-atmost-once-period-in-seconds="0" wrap-jdbc-objects="false">
     <property name="URL" value="jdbc:mysql://localhost:3306/wbd2?zeroDateTimeBehavior=convertToNull"/>
     <property name="User" value="root"/>
     <property name="Password" value=""/>
+    <property name="driverClass" value="com.mysql.jdbc.Driver"/>
   </jdbc-connection-pool>
 </resources>
